+++ src/geometry/mod.rs
@@ -9,7 +9,3 @@ pub mod constants;
 pub mod point;
 
 pub mod shapes;
-
-pub mod traits;
-
-pub mod geometry_relation;
