+++ src/geometry/shapes/mod.rs
@@ -2,4 +2,5 @@ pub mod line;
 pub mod polygon;
 pub mod solid;
 pub mod sphere;
+pub mod traits;
 pub mod triangle;
