+++ src/lib.rs
@@ -17,9 +17,8 @@ pub use geometry::shapes::polygon::Polygon;
 pub use geometry::shapes::solid::Solid;
 pub use geometry::shapes::sphere::Sphere;
 
-pub use geometry::geometry_relation::*;
+pub use geometry::shapes::traits::*;
 pub use geometry::shapes::triangle::Triangle;
-pub use geometry::traits::Geometry;
 pub use spatial_id::Block;
 pub use spatial_id::SpatialId;
 pub use spatial_id::collection::SpatialIdSet;
